Support for Decimals
Visible to User:
- clear in REPL should now work properly
- decimal values can now be assigned and printed

Internal:
- API Internal Error Warning Fix
- Doesn't break debugger on error (prints it out)
- Variables are now stored as Decimal instead of Int (allows division and subtraction)

+++ src/JukaCompiler/Expressions/Expr.cs
@@ -54,7 +54,7 @@ namespace JukaCompiler.Expressions
         }
         internal class Variable : Expr
         {
-            internal Int64 lexemeType;
+            internal Decimal lexemeType;
 
             internal Variable(Lexeme ExpressionLexeme)
             {
